first instance of full javascript test code
replaced login.js and script.js with code that is functional, index and login html files are completely temporary and can be overwritten, only note is that index.html is technically the login or welcome page since it is loaded first, so it must reference the login.js script, and then on the other hand, the login.html file will be the 8ball page, so it must reference the script.js script. Initial testing for storing users to local storage is commented out.